requirements.txt
Django>=5.0.0,<5.1
djangorestframework>=3.14.0,<3.15
django-cors-headers>=4.3.1,<4.4
djangorestframework-simplejwt>=5.3.1,<5.4
psycopg2-binary>=2.9.9,<3.0
drf_spectacular>=0.27.0,<1.0
PyJWT[crypto]>=2.8.0,<2.9
python-decouple>=3.8,<3.10
celery>=5.4.0,<5.5
redis>=5.0.8,<5.1
channels>=4.1.0,<5.0
daphne>=4.1.2,<5.0
channels_redis>=4.2.0,<5.0
twisted[http2,tls]
whitenoise>=6.8.2,<7.0
asyncio>=3.4.3,<4.0